优化异常处理器，改变了代码风格，所有异常不在Controller中捕捉，直接在Service或Facade层抛出直至统一异常处理器捕捉统一处理。

+++ cloud-gateway/cloud-gateway-core/src/main/java/com/lyl/study/cloud/gateway/core/service/impl/RoleServiceImpl.java
@@ -3,13 +3,14 @@ package com.lyl.study.cloud.gateway.core.service.impl;
 import com.baomidou.mybatisplus.service.impl.ServiceImpl;
 import com.lyl.study.cloud.base.exception.NoSuchEntityException;
 import com.lyl.study.cloud.base.idworker.Sequence;
-import com.lyl.study.cloud.gateway.api.dto.response.PermissionItem;
-import com.lyl.study.cloud.gateway.core.entity.Role;
-import com.lyl.study.cloud.gateway.core.mapper.RoleMapper;
-import com.lyl.study.cloud.gateway.core.service.RoleService;
+import com.lyl.study.cloud.gateway.api.GatewayErrorCode;
 import com.lyl.study.cloud.gateway.api.dto.request.RoleSaveForm;
 import com.lyl.study.cloud.gateway.api.dto.request.RoleUpdateForm;
-import com.lyl.study.cloud.gateway.api.dto.response.RoleDTO;
+import com.lyl.study.cloud.gateway.api.dto.response.PermissionItem;
+import com.lyl.study.cloud.gateway.core.entity.Role;
+import com.lyl.study.cloud.gateway.core.mapper.PermissionMapper;
+import com.lyl.study.cloud.gateway.core.mapper.RoleMapper;
+import com.lyl.study.cloud.gateway.core.service.RoleService;
 import org.springframework.beans.BeanUtils;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -33,6 +34,8 @@ import java.util.stream.Collectors;
 public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements RoleService {
     @Autowired
     private Sequence sequence;
+    @Autowired
+    private PermissionMapper permissionMapper;
 
     @Override
     @Transactional
@@ -46,12 +49,12 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
     }
 
     @Override
-    public void update(RoleUpdateForm form) {
+    public void update(RoleUpdateForm form) throws NoSuchEntityException {
         Long roleId = form.getId();
         Assert.notNull(roleId, "id cannot be null");
         Role record = baseMapper.selectById(roleId);
         if (record == null) {
-            throw new NoSuchEntityException("找不到ID为" + roleId + "的角色");
+            throw new NoSuchEntityException(GatewayErrorCode.NOT_FOUND, "找不到ID为" + roleId + "的角色");
         }
 
         BeanUtils.copyProperties(form, record);
@@ -64,6 +67,14 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
         Set<Long> permIdSet = permissions.stream().map(PermissionItem::getId).collect(Collectors.toSet());
         HashSet<Long> formPermIdSet = new HashSet<>(form.getPermissions());
         if (!permIdSet.equals(formPermIdSet)) {
+            // 校验授权项是否存在
+//            if (!formPermIdSet.isEmpty()) {
+//                List<Permission> records = permissionMapper.selectBatchIds(formPermIdSet);
+//                if (records.size() != formPermIdSet.size()) {
+//                    throw new NoSuchDependentedEntityException("找不到相关授权项");
+//                }
+//            }
+
             baseMapper.deleteRolePermissionsByRoleId(roleId);
             baseMapper.insertRolePermissions(roleId, formPermIdSet);
         }
@@ -71,14 +82,17 @@ public class RoleServiceImpl extends ServiceImpl<RoleMapper, Role> implements Ro
 
     @Override
     @Transactional
-    public int deleteById(long id) {
+    public void deleteById(long id) throws NoSuchEntityException {
         // 删除角色记录
         int rows = baseMapper.deleteById(id);
-        // 删除角色授权关联关系
-        baseMapper.deleteRolePermissionsByRoleId(id);
-        // 删除用户角色关联关系
-        baseMapper.deleteUserRolesByRoleId(id);
-        return rows;
+        if (rows != 0) {
+            // 删除角色授权关联关系
+            baseMapper.deleteRolePermissionsByRoleId(id);
+            // 删除用户角色关联关系
+            baseMapper.deleteUserRolesByRoleId(id);
+        } else {
+            throw new NoSuchEntityException("找不到ID为" + id + "的角色");
+        }
     }
 
     @Override
